Feat config refactor (#51)
* save

* adjust ipfs service

+++ pkg/im/global_context.go
@@ -15,7 +15,7 @@ var (
 	CurrentMilestoneIndex          = uint32(0)
 	CurrentMilestoneTimestamp      = uint32(0)
 	LastTimeReceiveEventFromHornet = uint32(0)
-	HornetChainId                  = 0
+	HornetChainId                  = uint32(0)
 	HornetChainName                = ""
 	CurrentNetwork                 = ShimmerMainNet
 	NodeHTTPAPIClient              *nodeclient.Client
